Music converter base functionality (#2)
This pull request includes the base functionality for interacting with Google Music, YouTube, and Spotify. This logic is not yet hooked up to Telegram. Calls to the telegram/musicConverter endpoint return a 400 error.

A number of changes needed to be made to the development environment to get the gmusic api package working well with amazonlinux. After many days of experimenting with different options, this PR:

 * replaces pyenv with pipenv for dependency management
 * dockerizes builds with serverless-python-requirements to make sure packages are compatible with amazonlinux

This PR should be in a state where these features can be pushed without impacting the existing alert bot.